# Speed up querying for unencrypted builds 