--- a/atc/db/migration/migrations/1595347367_add_index_for_unencrypted_private_plans_to_builds.down.sql
+++ b/atc/db/migration/migrations/1595347367_add_index_for_unencrypted_private_plans_to_builds.down.sql
@@ -1,0 +1,3 @@
+BEGIN;
+  DROP INDEX unencrypted_private_plans_build_idx;
+COMMIT;