--- a/atc/db/migration/migrations/1595347367_add_index_for_unencrypted_private_plans_to_builds.up.sql
+++ b/atc/db/migration/migrations/1595347367_add_index_for_unencrypted_private_plans_to_builds.up.sql
@@ -1,0 +1,3 @@
+BEGIN;
+  CREATE INDEX unencrypted_private_plans_build_idx ON builds (id, private_plan) WHERE nonce IS NULL AND private_plan IS NOT NULL;
+COMMIT;